Fix incorrect input type validation over nullable array items
