+++ amf-cli/shared/src/test/resources/graphql/tck/apis/invalid/arguments-list.api.federation.report
@@ -1,33 +1,9 @@
 ModelId: file://amf-cli/shared/src/test/resources/graphql/tck/apis/invalid/arguments-list.api.graphql
 Profile: GraphQL
 Conforms: false
-Number of results: 10
+Number of results: 6
 
 Level: Violation
-
-- Constraint: http://a.ml/vocabularies/amf/parser#invalid-input-type
-  Message: Type '[String]' from field 'valid' must be an input type
-  Severity: Violation
-  Target: file://amf-cli/shared/src/test/resources/graphql/tck/apis/invalid/arguments-list.api.graphql#/declares/valid1/shape/default-node
-  Property: http://www.w3.org/ns/shacl#property
-  Range: [(6,18)-(6,46)]
-  Location: 
-
-- Constraint: http://a.ml/vocabularies/amf/parser#invalid-input-type
-  Message: Type '[[String]]' from field 'valid' must be an input type
-  Severity: Violation
-  Target: file://amf-cli/shared/src/test/resources/graphql/tck/apis/invalid/arguments-list.api.graphql#/declares/valid2/shape/default-node
-  Property: http://www.w3.org/ns/shacl#property
-  Range: [(7,18)-(7,62)]
-  Location: 
-
-- Constraint: http://a.ml/vocabularies/amf/parser#invalid-input-type
-  Message: Type '[Boolean]' from field 'invalid' must be an input type
-  Severity: Violation
-  Target: file://amf-cli/shared/src/test/resources/graphql/tck/apis/invalid/arguments-list.api.graphql#/declares/invalid1/shape/default-node
-  Property: http://www.w3.org/ns/shacl#property
-  Range: [(8,20)-(8,51)]
-  Location: 
 
 - Constraint: http://a.ml/vocabularies/amf/parser#invalid-default-value-type-directive
   Message: Expecting 'Boolean' value but got 'String' value instead
@@ -43,14 +19,6 @@ Level: Violation
   Target: file://amf-cli/shared/src/test/resources/graphql/tck/apis/invalid/arguments-list.api.graphql#/declares/invalid1/shape/default-node
   Property: http://www.w3.org/ns/shacl#defaultValue
   Range: [(8,47)-(8,50)]
-  Location: 
-
-- Constraint: http://a.ml/vocabularies/amf/parser#invalid-input-type
-  Message: Type '[[Boolean]]' from field 'invalid' must be an input type
-  Severity: Violation
-  Target: file://amf-cli/shared/src/test/resources/graphql/tck/apis/invalid/arguments-list.api.graphql#/declares/invalid2/shape/default-node
-  Property: http://www.w3.org/ns/shacl#property
-  Range: [(9,20)-(9,67)]
   Location: 
 
 - Constraint: http://a.ml/vocabularies/amf/parser#invalid-default-value-type-directive
